python/weatherapp-flask/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
requests2==2.16.0